fix: QA issues - fix broken tests, add thread safety, error handling, and code quality improvements
- Fix namespace errors in Application.Tests (Domain.Order→Domain.Orders, Domain.Payment→Domain.Payments)
- Add Application.Tests to solution file
- Replace Dictionary with ConcurrentDictionary in InMemoryOrderRepository
- Add error handling for domain validation (return 400 instead of 500)
- Throw on missing items in Order.RemoveItem/ChangeItemQuantity/ChangeItemPrice
- Fix OrderItemDto.Price type from decimal to long (matches Money.Amount)
- Validate Money.Currency is not null/empty
- Remove orphan root Program.cs
- Expose OrderItems as IReadOnlyList<Item>
- Remove unused FluentValidation dependency and registration
- Remove unused using directives

Agent-Logs-Url: https://github.com/KotBehamot/ShopExample/sessions/879b4aa6-afdb-4873-964a-ba7e8ac16e10

+++ Application/Orders/Queries/GetOrder/OrderItemDto.cs
@@ -2,5 +2,5 @@
 
 namespace Application.Orders.Queries.GetOrder
 {
-    public readonly record struct OrderItemDto(Guid Id, string Name, int Quantity, decimal Price);
+    public readonly record struct OrderItemDto(Guid Id, string Name, int Quantity, long Price);
 }
